+++ CHANGES.rst
@@ -6,6 +6,8 @@ Changelog
 
 - execute command accepts kwargs now
 
+- execute command returns the command value now
+
 
 0.3.1 (2018-01-04)
 ==================